refactor: log 서비스 보완
