+++ splanet.log
@@ -1,3 +1,4 @@
 eventType: LOGIN_SUCCESS, userId: 3718722437, deviceId: 3718722437, timestamp: 2024-11-08T20:41:53.096336
 eventType: LOGIN_SUCCESS, userId: 3718722437, deviceId: 3718722437, timestamp: 2024-11-08T20:42:48.271852
 eventType: LOGIN_SUCCESS, userId: 3718722437, deviceId: 3718722437, timestamp: 2024-11-08T20:42:58.578322
+eventType: LOGIN_SUCCESS, userId: 1, deviceId: d3844005-40dc-4441-93ec-a4cfedebc1b7, timestamp: 2024-11-10T21:44:54.248766, requestPath: /login/oauth2/code/kakao, headers: host: localhost:8080, connection: keep-alive, upgrade-insecure-requests: 1, user-agent: Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/130.0.0.0 Safari/537.36, accept: text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.7, sec-fetch-site: cross-site, sec-fetch-mode: navigate, sec-fetch-user: ?1, sec-fetch-dest: document, sec-ch-ua: "Chromium";v="130", "Google Chrome";v="130", "Not?A_Brand";v="99", sec-ch-ua-mobile: ?0, sec-ch-ua-platform: "macOS", referer: http://localhost:8080/, accept-encoding: gzip, deflate, br, zstd, accept-language: ko-KR,ko;q=0.9,en-US;q=0.8,en;q=0.7, cookie: Idea-559f79a3=fc422438-a915-454a-9c16-e8ea96f7b5fc; grafana_session=d9ebba25f7765c76302cfdbb4dee6911; grafana_session_expiry=1731049627; JSESSIONID=E8141687EC1FA179C28E4E314369975D,
