+++ src/main/java/com/splanet/splanet/oauth/OAuth2AuthenticationSuccessHandler.java
@@ -3,6 +3,7 @@ package com.splanet.splanet.oauth;
 import com.splanet.splanet.core.properties.OAuth2Properties;
 import com.splanet.splanet.jwt.JwtTokenProvider;
 import com.splanet.splanet.jwt.service.TokenService;
+import com.splanet.splanet.log.service.LogService; // LogService 추가
 import com.splanet.splanet.user.entity.User;
 import com.splanet.splanet.user.repository.UserRepository;
 import jakarta.servlet.ServletException;
@@ -22,19 +23,18 @@ import java.util.UUID;
 public class OAuth2AuthenticationSuccessHandler implements AuthenticationSuccessHandler {
 
   private final JwtTokenProvider jwtTokenProvider;
-
   private final UserRepository userRepository;
   private final TokenService tokenService;
   private final OAuth2Properties oAuth2Properties;
+  private final LogService logService; // LogService 필드 추가
 
-  public OAuth2AuthenticationSuccessHandler(JwtTokenProvider jwtTokenProvider, UserRepository userRepository, TokenService tokenService, OAuth2Properties oAuth2Properties) {
+  public OAuth2AuthenticationSuccessHandler(JwtTokenProvider jwtTokenProvider, UserRepository userRepository, TokenService tokenService, OAuth2Properties oAuth2Properties, LogService logService) {
     this.jwtTokenProvider = jwtTokenProvider;
     this.userRepository = userRepository;
     this.tokenService = tokenService;
     this.oAuth2Properties = oAuth2Properties;
+    this.logService = logService; // LogService 초기화
   }
-
-
 
   @Override
   public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
@@ -66,6 +66,10 @@ public class OAuth2AuthenticationSuccessHandler implements AuthenticationSuccess
 
     tokenService.storeRefreshToken(refreshToken, user.getId(), deviceId);
 
+    // 로그인 성공 시 로그 기록
+    String requestPath = request.getRequestURI();
+    String headers = getHeadersAsString(request);
+    logService.recordLoginLog(user.getId(), deviceId, requestPath, headers);
 
     String redirectUrlWithParams = UriComponentsBuilder.fromUriString(oAuth2Properties.getRedirectDevUrl())
             .queryParam("access", accessToken)
@@ -82,5 +86,12 @@ public class OAuth2AuthenticationSuccessHandler implements AuthenticationSuccess
     return userRepository.findByNickname(nickname)
             .map(existingUser -> nickname + "#" + uniqueSuffix)
             .orElse(nickname);
+  }
+
+  private String getHeadersAsString(HttpServletRequest request) {
+    StringBuilder headers = new StringBuilder();
+    request.getHeaderNames().asIterator().forEachRemaining(headerName ->
+            headers.append(headerName).append(": ").append(request.getHeader(headerName)).append(", "));
+    return headers.toString();
   }
 }
